Moved BarcodeResult to StepResult
