+++ MWCameraPlugin/MWCameraPlugin/BarcodeStep/MWBarcodeStepViewController.swift
@@ -17,7 +17,7 @@ public class MWBarcodeStepViewController: MWStepViewController {
         super.viewDidLoad()
         let barcodeScanner = _MWBarcodeStepViewController() { [weak self] codeFound in
             guard let self = self else { return }
-            let result = MWBarcodeResult(identifier: self.barcodeStep.identifier, codeFound: codeFound)
+            let result = BarcodeResult(identifier: self.barcodeStep.identifier, codeFound: codeFound)
             self.addStepResult(result)
             self.goForward()
         }
